// Modules/Angular/Day11-12/Disha-Shah/Assignment/StudentDynamicFormApp/src/app/models/student-base.ts
export class StudentBase<T> {
    value?: T;
    key: string;
    label: string;
    required: boolean;
    order: number;
    controlType: string;
    type: string;
  
    constructor(options: {
        value?: T;
        key?: string;
        label?: string;
        required?: boolean;
        order?: number;
        controlType?: string;
        type?: string;
      } = {}) {
      this.value = options.value;
      this.key = options.key || '';
      this.label = options.label || '';
      this.required = !!options.required;
      this.order = options.order === undefined ? 1 : options.order;
      this.controlType = options.controlType || '';
      this.type = options.type || '';
    }
}